refactor out activity list
