+++ apps/admin_panel/assets/src/omg-transaction-request-tab/index.js
@@ -2,22 +2,19 @@ import React, { Component } from 'react'
 import PropTypes from 'prop-types'
 import styled from 'styled-components'
 import TabPanel from './TabPanel'
-import SortableTable from '../omg-table'
-import ConsumptionFetcherByTransactionIdFetcher from '../omg-consumption/consumptionByTransactionIdFetcher'
 import TransactionRequestProvider from '../omg-transaction-request/transactionRequestProvider'
 import { Icon, Button, Select, Input } from '../omg-uikit'
 import { withRouter } from 'react-router-dom'
 import queryString from 'query-string'
 import QR from './QrCode'
-import moment from 'moment'
 import { connect } from 'react-redux'
-import { approveConsumptionById, rejectConsumptionById } from '../omg-consumption/action'
 import { compose } from 'recompose'
 import { formatNumber } from '../utils/formatter'
 import AllWalletsFetcher from '../omg-wallet/allWalletsFetcher'
 import TokensFetcher from '../omg-token/tokensFetcher'
 import { consumeTransactionRequest } from '../omg-transaction-request/action'
 import { selectGetTransactionRequestById } from '../omg-transaction-request/selector'
+import ActivityList from './ActivityList'
 const PanelContainer = styled.div`
   height: 100vh;
   position: fixed;
@@ -35,32 +32,7 @@ const PanelContainer = styled.div`
     padding: 20px;
   }
 `
-const ContentContainer = styled.div`
-  height: calc(100vh - 160px);
-  overflow: auto;
-  table tr td {
-    height: 22px;
-    vertical-align: middle;
-  }
-  td:first-child {
-    max-width: 100px;
-    width: 100px;
-    > div {
-      white-space: nowrap;
-      overflow: hidden;
-      text-overflow: ellipsis;
-    }
-  }
-  td:nth-child(2) {
-    max-width: 120px;
-    width: 120px;
-    > div {
-      white-space: nowrap;
-      overflow: hidden;
-      text-overflow: ellipsis;
-    }
-  }
-`
+
 const ConsumeActionContainer = styled.form`
   display: flex;
   margin: 20px 0;
@@ -129,18 +101,7 @@ const AdditionalRequestDataContainer = styled.div`
     letter-spacing: 1px;
   }
 `
-const ConfirmButton = styled.button`
-  display: inline-block;
-  background-color: white;
-  cursor: pointer;
-  border-radius: 2px;
-  padding: 4px 10px;
-  line-height: 10px;
-  color: ${props => props.theme.colors.B200};
-  i {
-    font-size: 10px;
-  }
-`
+
 const QrTypeContainer = styled.div`
   margin-top: 15px;
   :not(:last-child) {
@@ -150,18 +111,7 @@ const QrTypeContainer = styled.div`
     color: ${props => props.theme.colors.B300};
   }
 `
-const ConfirmButtonApprove = ConfirmButton.extend`
-  :hover {
-    color: #16826b;
-    background-color: #dcfaf4;
-  }
-`
-const ConfirmButtonReject = ConfirmButton.extend`
-  :hover {
-    color: #d51404;
-    background-color: #ffefed;
-  }
-`
+
 const InputLabel = styled.div`
   margin-top: 20px;
   font-size: 14px;
@@ -185,7 +135,7 @@ const enhance = compose(
     state => ({
       selectTransactionRequestById: selectGetTransactionRequestById(state)
     }),
-    { approveConsumptionById, rejectConsumptionById, consumeTransactionRequest }
+    { consumeTransactionRequest }
   )
 )
 const Error = styled.div`
@@ -201,8 +151,6 @@ class TransactionRequestPanel extends Component {
   static propTypes = {
     history: PropTypes.object,
     location: PropTypes.object,
-    rejectConsumptionById: PropTypes.func,
-    approveConsumptionById: PropTypes.func,
     consumeTransactionRequest: PropTypes.func,
     selectTransactionRequestById: PropTypes.func
   }
@@ -224,21 +172,8 @@ class TransactionRequestPanel extends Component {
   constructor (props) {
     super(props)
     this.state = {}
-    this.columns = [
-      { key: 'amount', title: 'AMOUNT', sort: true },
-      { key: 'to', title: 'TO' },
-      { key: 'created_at', title: 'CREATED DATE', sort: true },
-      { key: 'status', title: 'CONFIRMATION' }
-    ]
   }
-  onClickConfirm = id => async e => {
-    e.stopPropagation()
-    await this.props.approveConsumptionById(id)
-  }
-  onClickReject = id => e => {
-    e.stopPropagation()
-    this.props.rejectConsumptionById(id)
-  }
+
   onChangeWalletInput = e => {
     this.setState({ consumeAddress: e.target.value })
   }
@@ -254,15 +189,7 @@ class TransactionRequestPanel extends Component {
   onSelectTokenSelect = token => {
     this.setState({ searchTokenValue: token.value, selectedToken: token })
   }
-  onClickRow = (data, index) => e => {
-    const searchObject = queryString.parse(this.props.location.search)
-    this.props.history.push({
-      search: queryString.stringify({
-        ...searchObject,
-        [`show-consumption-tab`]: data.id
-      })
-    })
-  }
+
   onClickClose = () => {
     const searchObject = queryString.parse(this.props.location.search)
     delete searchObject['active-tab']
@@ -302,50 +229,7 @@ class TransactionRequestPanel extends Component {
       this.setState({ submitStatus: 'FAILED', error: `${error}` })
     }
   }
-  rowRenderer = (key, data, rows) => {
-    if (key === 'amount') {
-      return (
-        <div>
-          {formatNumber((data || 0) / _.get(rows, 'token.subunit_to_unit'))}{' '}
-          {_.get(rows, 'token.symbol')}
-        </div>
-      )
-    }
-    if (key === 'created_at') {
-      return moment(data).format('DD/MM hh:mm:ss')
-    }
-    if (key === 'to') {
-      return <div>{rows.user_id || _.get(rows, 'account.name')}</div>
-    }
 
-    if (key === 'status') {
-      switch (data) {
-        case 'pending':
-          return (
-            <div>
-              <ConfirmButtonApprove onClick={this.onClickConfirm(rows.id)}>
-                <Icon name='Checked' />
-              </ConfirmButtonApprove>
-              <ConfirmButtonReject onClick={this.onClickReject(rows.id)}>
-                <Icon name='Close' />
-              </ConfirmButtonReject>
-            </div>
-          )
-        case 'confirmed': {
-          return 'Confirmed'
-        }
-        case 'failed': {
-          return 'Failed'
-        }
-        case 'rejected': {
-          return 'Rejected'
-        }
-        default:
-          return <span>{data}</span>
-      }
-    }
-    return data
-  }
   getExpiredReason = reason => {
     switch (reason) {
       case 'max_consumptions_reached':
@@ -356,35 +240,7 @@ class TransactionRequestPanel extends Component {
         return 'Expired.'
     }
   }
-  renderActivityList = () => {
-    return (
-      <ConsumptionFetcherByTransactionIdFetcher
-        id={queryString.parse(this.props.location.search)['show-request-tab']}
-        render={({ data, individualLoadingStatus, pagination }) => {
-          return (
-            <ContentContainer>
-              <SortableTable
-                rows={data}
-                columns={this.columns}
-                loadingStatus={individualLoadingStatus}
-                rowRenderer={this.rowRenderer}
-                onClickRow={this.onClickRow}
-                isFirstPage={pagination.is_first_page}
-                isLastPage={pagination.is_last_page}
-                navigation
-                pageEntity={'page-activity'}
-              />
-            </ContentContainer>
-          )
-        }}
-        query={{
-          page: queryString.parse(this.props.location.search)['page-activity'],
-          perPage: Math.floor(window.innerHeight / 65),
-          uniqueId: queryString.parse(this.props.location.search)['show-request-tab']
-        }}
-      />
-    )
-  }
+
   renderProperties = transactionRequest => {
     const valid = transactionRequest.status === 'valid'
     return (
@@ -541,7 +397,7 @@ class TransactionRequestPanel extends Component {
                   {
                     key: 'activity',
                     tabTitle: 'ACTIVITY LIST',
-                    tabContent: this.renderActivityList()
+                    tabContent: <ActivityList />
                   },
                   {
                     key: 'properties',
